+++ src/main/java/de/alpharogroup/collections/properties/PropertiesExtensions.java
@@ -177,15 +177,39 @@ public final class PropertiesExtensions
 	 */
 	public static Properties loadProperties(final File propertiesFile) throws IOException
 	{
+		return loadProperties(propertiesFile, false);
+	}
+
+	/**
+	 * Load a Properties-object from the given File-object.
+	 *
+	 * @param propertiesFile
+	 *            the properties file
+	 * @param loadFromXML
+	 *            the load from XML
+	 * @return the properties or null if the file is not found.
+	 * @throws IOException
+	 *             Signals that an I/O exception has occurred.
+	 */
+	public static Properties loadProperties(final File propertiesFile, boolean loadFromXML)
+		throws IOException
+	{
 		Properties properties = null;
-		InputStream is = null;
+		InputStream inputStream = null;
 		if (propertiesFile.exists())
 		{
-			is = propertiesFile.toURI().toURL().openStream();
-			if (is != null)
+			inputStream = propertiesFile.toURI().toURL().openStream();
+			if (inputStream != null)
 			{
 				properties = new Properties();
-				properties.load(is);
+				if (loadFromXML)
+				{
+					properties.loadFromXML(inputStream);
+				}
+				else
+				{
+					properties.load(inputStream);
+				}
 			}
 		}
 		else
@@ -240,9 +264,143 @@ public final class PropertiesExtensions
 	public static void toProperties(final OutputStream properties, final InputStream xml,
 		final String comment) throws IOException
 	{
-		final Properties prop = new Properties();
-		prop.loadFromXML(xml);
-		prop.store(properties, comment);
+		toProperties(properties, xml, comment, true);
+	}
+
+	/**
+	 * Converts the given properties InputStream to the given properties {@link OutputStream} that
+	 * represents the output file.
+	 *
+	 * @param outputStream
+	 *            the properties file. The xml file does not have to exist.
+	 * @param inputStream
+	 *            the xml file with the properties to convert.
+	 * @param comment
+	 *            the comment
+	 * @throws IOException
+	 *             Signals that an I/O exception has occurred.
+	 */
+	public static void toPropertiesFile(final OutputStream outputStream,
+		final InputStream inputStream, final String comment) throws IOException
+	{
+		toProperties(outputStream, inputStream, comment, false);
+	}
+
+	/**
+	 * Exports the given {@link InputStream} that represents a properties file to the given
+	 * properties {@link OutputStream} that represents the output file. The flag xmlFile tells if
+	 * the output shell be an xml file or a properties file.
+	 *
+	 * @param outputStream
+	 *            the stream from the output properties file. If the file does not exists a new file
+	 *            will be created.
+	 * @param inputStream
+	 *            the stream from input file with the properties to convert
+	 * @param comment
+	 *            the comment
+	 * @param xmlFile
+	 *            the flag that tells if the output shell be an xml file if true otherwise it will
+	 *            be an properties file
+	 * @throws IOException
+	 *             Signals that an I/O exception has occurred.
+	 */
+	public static void toProperties(final OutputStream outputStream, final InputStream inputStream,
+		final String comment, boolean xmlFile) throws IOException
+	{
+		export(null, outputStream, inputStream, comment, xmlFile, false);
+	}
+
+	/**
+	 * Exports the given {@link InputStream} that represents a properties file to the given
+	 * properties {@link OutputStream} that represents the output file. The flag xmlFile tells if
+	 * the output shell be an xml file or a properties file.
+	 *
+	 * @param properties
+	 *            the properties to store
+	 * @param outputStream
+	 *            the stream from the output properties file. If the file does not exists a new file
+	 *            will be created.
+	 * @param inputStream
+	 *            the input stream
+	 * @param comment
+	 *            the comment
+	 * @param loadFromXML
+	 *            the flag that tells if the input shell be loaded from a XML file if true otherwise
+	 *            it will be loaded from a properties file
+	 * @param storeToXML
+	 *            the flag that tells if the output shell be stored to a XML file if true otherwise
+	 *            it will be stored to a properties file
+	 * @throws IOException
+	 *             Signals that an I/O exception has occurred.
+	 */
+	public static void export(Properties properties, OutputStream outputStream,
+		InputStream inputStream, String comment, boolean loadFromXML, boolean storeToXML)
+		throws IOException
+	{ // create a new properties if given properties object is null
+		if (properties == null)
+		{
+			properties = new Properties();
+		}
+		// load from input stream if not null
+		if (inputStream != null)
+		{
+			if (loadFromXML)
+			{
+				properties.loadFromXML(inputStream);
+			}
+			else
+			{
+				properties.load(inputStream);
+			}
+		}
+		// store as xml or as normal properties format
+		if (storeToXML)
+		{
+			properties.storeToXML(outputStream, comment);
+		}
+		else
+		{
+			properties.store(outputStream, comment);
+		}
+	}
+
+	/**
+	 * Exports the given {@link InputStream} that represents a properties file to the given
+	 * properties {@link OutputStream} that represents the output file. The flag xmlFile tells if
+	 * the output shell be an xml file or a properties file.
+	 *
+	 * @param properties
+	 *            the properties to store
+	 * @param outputStream
+	 *            the stream from the output properties file. If the file does not exists a new file
+	 *            will be created.
+	 * @throws IOException
+	 *             Signals that an I/O exception has occurred.
+	 */
+	public static void export(Properties properties, OutputStream outputStream) throws IOException
+	{
+		export(properties, outputStream, null);
+	}
+
+	/**
+	 * Exports the given {@link InputStream} that represents a properties file to the given
+	 * properties {@link OutputStream} that represents the output file. The flag xmlFile tells if
+	 * the output shell be an xml file or a properties file.
+	 *
+	 * @param properties
+	 *            the properties to store
+	 * @param outputStream
+	 *            the stream from the output properties file. If the file does not exists a new file
+	 *            will be created.
+	 * @param comment
+	 *            the comment
+	 * @throws IOException
+	 *             Signals that an I/O exception has occurred.
+	 */
+	public static void export(Properties properties, OutputStream outputStream, String comment)
+		throws IOException
+	{
+		export(properties, outputStream, null, comment, false, false);
 	}
 
 	/**